src/python/requirements.txt
google-adk
openai
pydantic
langchain
anthropic
mcp
python-dotenv
fastapi
uvicorn[standard]
httpx